requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-environ==0.11.2
Pillow==10.4.0
